Factor out jxl_cms as cmake library
This allows us to use GetJxlCms from the decoder
