+++ lib/extras/packed_image_convert.cc
@@ -14,9 +14,9 @@
 #include "lib/jxl/color_encoding_internal.h"
 #include "lib/jxl/color_management.h"
 #include "lib/jxl/dec_external_image.h"
-#include "lib/jxl/enc_color_management.h"
 #include "lib/jxl/enc_external_image.h"
 #include "lib/jxl/enc_image_bundle.h"
+#include "lib/jxl/jxl_cms.h"
 #include "lib/jxl/luminance.h"
 
 namespace jxl {
