+++ lib/extras/tone_mapping.cc
@@ -11,8 +11,8 @@
 #include <hwy/highway.h>
 
 #include "lib/jxl/dec_tone_mapping-inl.h"
-#include "lib/jxl/enc_color_management.h"
 #include "lib/jxl/image_bundle.h"
+#include "lib/jxl/jxl_cms.h"
 
 HWY_BEFORE_NAMESPACE();
 namespace jxl {
