+++ lib/jxl/enc_xyb.cc
@@ -20,12 +20,12 @@
 #include "lib/jxl/color_encoding_internal.h"
 #include "lib/jxl/color_management.h"
 #include "lib/jxl/enc_bit_writer.h"
-#include "lib/jxl/enc_color_management.h"
 #include "lib/jxl/enc_image_bundle.h"
 #include "lib/jxl/fast_math-inl.h"
 #include "lib/jxl/fields.h"
 #include "lib/jxl/image_bundle.h"
 #include "lib/jxl/image_ops.h"
+#include "lib/jxl/jxl_cms.h"
 #include "lib/jxl/opsin_params.h"
 #include "lib/jxl/transfer_functions-inl.h"
