+++ lib/jxl_lists.bzl
@@ -30,6 +30,12 @@ libjxl_base_sources = [
     "jxl/base/status.h",
 ]
 
+libjxl_cms_sources = [
+    "jxl/jxl_cms.cc",
+    "jxl/jxl_cms.h",
+    "jxl/jxl_skcms.h",
+]
+
 libjxl_codec_apng_sources = [
     "extras/dec/apng.cc",
     "extras/dec/apng.h",
@@ -323,8 +329,6 @@ libjxl_enc_sources = [
     "jxl/enc_cluster.h",
     "jxl/enc_coeff_order.cc",
     "jxl/enc_coeff_order.h",
-    "jxl/enc_color_management.cc",
-    "jxl/enc_color_management.h",
     "jxl/enc_comparator.cc",
     "jxl/enc_comparator.h",
     "jxl/enc_context_map.cc",
@@ -362,7 +366,6 @@ libjxl_enc_sources = [
     "jxl/enc_icc_codec.h",
     "jxl/enc_image_bundle.cc",
     "jxl/enc_image_bundle.h",
-    "jxl/enc_jxl_skcms.h",
     "jxl/enc_linalg.cc",
     "jxl/enc_linalg.h",
     "jxl/enc_modular.cc",
