+++ tools/butteraugli_main.cc
@@ -22,10 +22,10 @@
 #include "lib/jxl/color_encoding_internal.h"
 #include "lib/jxl/color_management.h"
 #include "lib/jxl/enc_butteraugli_comparator.h"
-#include "lib/jxl/enc_color_management.h"
 #include "lib/jxl/image.h"
 #include "lib/jxl/image_bundle.h"
 #include "lib/jxl/image_ops.h"
+#include "lib/jxl/jxl_cms.h"
 #include "tools/file_io.h"
 #include "tools/thread_pool_internal.h"
