+++ tools/comparison_viewer/image_loading.cc
@@ -10,9 +10,9 @@
 
 #include "lib/extras/codec.h"
 #include "lib/extras/dec/color_hints.h"
-#include "lib/jxl/enc_color_management.h"
 #include "lib/jxl/image_bundle.h"
 #include "lib/jxl/image_metadata.h"
+#include "lib/jxl/jxl_cms.h"
 #include "tools/file_io.h"
 #include "tools/thread_pool_internal.h"
 #include "tools/viewer/load_jxl.h"
