+++ tools/djxl_fuzzer_corpus.cc
@@ -30,12 +30,12 @@
 #include "lib/jxl/enc_ans.h"
 #include "lib/jxl/enc_aux_out.h"
 #include "lib/jxl/enc_cache.h"
-#include "lib/jxl/enc_color_management.h"
 #include "lib/jxl/enc_external_image.h"
 #include "lib/jxl/enc_file.h"
 #include "lib/jxl/enc_params.h"
 #include "lib/jxl/encode_internal.h"
 #include "lib/jxl/jpeg/enc_jpeg_data.h"
+#include "lib/jxl/jxl_cms.h"
 #include "lib/jxl/modular/encoding/context_predict.h"
 #include "tools/file_io.h"
 #include "tools/thread_pool_internal.h"
