+++ tools/hdr/exr_to_pq.cc
@@ -9,8 +9,8 @@
 #include "lib/extras/codec.h"
 #include "lib/extras/dec/decode.h"
 #include "lib/extras/packed_image_convert.h"
-#include "lib/jxl/enc_color_management.h"
 #include "lib/jxl/image_bundle.h"
+#include "lib/jxl/jxl_cms.h"
 #include "tools/cmdline.h"
 #include "tools/file_io.h"
 #include "tools/hdr/image_utils.h"
