+++ tools/hdr/pq_to_hlg.cc
@@ -9,7 +9,7 @@
 #include "lib/extras/codec.h"
 #include "lib/extras/hlg.h"
 #include "lib/extras/tone_mapping.h"
-#include "lib/jxl/enc_color_management.h"
+#include "lib/jxl/jxl_cms.h"
 #include "tools/args.h"
 #include "tools/cmdline.h"
 #include "tools/hdr/image_utils.h"
