+++ tools/ssimulacra_main.cc
@@ -7,8 +7,8 @@
 
 #include "lib/extras/codec.h"
 #include "lib/jxl/color_management.h"
-#include "lib/jxl/enc_color_management.h"
 #include "lib/jxl/image_bundle.h"
+#include "lib/jxl/jxl_cms.h"
 #include "tools/file_io.h"
 #include "tools/ssimulacra.h"
